@@ -8,45 +8,52 @@
     "start": "next start",
     "lint": "next lint",
     "lint-fix": "next lint --fix --ext .js,.jsx .",
-    "postbuild": "next-sitemap --config next-sitemap.config.js"
+    "postbuild": "next-sitemap --config next-sitemap.config.js",
+    "analyze": "ANALYZE=true next build"
   },
   "dependencies": {
-    "@portabletext/react": "^3.0.15",
-    "@reduxjs/toolkit": "^2.2.1",
-    "@sanity/code-input": "^4.1.3",
-    "@sanity/react-loader": "^1.9.5",
-    "@sanity/vision": "^3.36.0",
-    "@superfaceai/one-sdk": "^2.4.2",
-    "dayjs": "^1.11.10",
-    "framer-motion": "^10.6.0",
-    "next": "^14.1.0",
-    "next-sanity": "^8.5.0",
-    "npm": "^10.5.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-redux": "^9.1.0",
-    "react-syntax-highlighter": "^15.5.0",
+    "@portabletext/react": "^4.0.1",
+    "@reduxjs/toolkit": "^2.8.2",
+    "@sanity/code-input": "^6.0.1",
+    "@sanity/react-loader": "^1.11.18",
+    "@sanity/vision": "^4.5.0",
+    "@sendgrid/mail": "^8.1.5",
+    "@superfaceai/one-sdk": "^2.5.0",
+    "@tsparticles/react": "^3.0.0",
+    "dayjs": "^1.11.13",
+    "framer-motion": "^12.23.12",
+    "next": "^15.5.0",
+    "next-sanity": "^10.0.13",
+    "nodemailer": "^7.0.5",
+    "npm": "^11.5.2",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-redux": "^9.2.0",
+    "react-syntax-highlighter": "^15.6.3",
     "react-tsparticles": "^2.12.2",
-    "sanity": "^3.36.0",
-    "sanity-plugin-iframe-pane": "^3.1.6",
-    "sharp": "^0.32.6",
-    "styled-components": "^6.0.7",
+    "sanity": "^4.5.0",
+    "sanity-plugin-iframe-pane": "^4.0.0",
+    "sharp": "^0.34.3",
+    "styled-components": "^6.1.19",
     "tsparticles": "^2.12.0"
   },
+  "engines": {
+    "node": ">=23.0.0"
+  },
   "devDependencies": {
-    "@sanity/eslint-config-studio": "^3.0.1",
-    "@tailwindcss/typography": "^0.5.12",
-    "autoprefixer": "^10.4.15",
-    "eslint": "^8.48.0",
-    "eslint-config-next": "^14.1.4",
-    "eslint-config-prettier": "^9.0.0",
+    "@sanity/eslint-config-studio": "^5.0.2",
+    "@tailwindcss/typography": "^0.5.16",
+    "autoprefixer": "^10.4.21",
+    "eslint": "^9.34.0",
+    "eslint-config-next": "^15.5.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-next": "^0.0.0",
-    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react": "^7.37.5",
     "next-sitemap": "^4.2.3",
-    "postcss": "^8.4.29",
-    "tailwindcss": "^3.3.3",
-    "typescript": "^5.2.2"
+    "postcss": "^8.5.6",
+    "tailwindcss": "^3.4.17",
+    "typescript": "^5.9.2"
   }
 }
