@@ -1,16 +1,15 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     reactStrictMode: true,
-    swcMinify: true,
     images: {
       remotePatterns: [
         {
-        protocol: "https",
-        hostname: "cdn.sanity.io",
-        port:""
+          protocol: "https",
+          hostname: "cdn.sanity.io",
+          port: ""
         }
-      ],
-  }
+      ]
+    }
 }
-  
-  module.exports = nextConfig
+
+module.exports = nextConfig
