@@ -1,5 +1,10 @@
+export const metadata = {
+  title: 'Next.js',
+  description: 'Generated by Next.js',
+}
+
 export default function RootLayout({ children }) {
- return (
+  return (
     <html lang="en">
       <body>{children}</body>
     </html>
